requirements-cuda.txt
# =========================
# Web framework
# =========================
fastapi==0.95.2
uvicorn==0.22.0
gunicorn==23.0.0
httptools==0.5.0
watchgod==0.6
websockets==10.4
python-multipart>=0.0.6,<0.1.0
aiofiles==0.8.0
requests==2.28.2

# =========================
# Data processing
# =========================
pydantic==1.10.8
numpy==1.24.4
pandas>=1.5.0,<2.0.0
tqdm==4.64.1

# =========================
# Excel handling
# =========================
openpyxl==3.0.10

# =========================
# Image processing
# =========================
opencv-python==4.6.0.66
Pillow==9.5.0
scikit-image==0.20.0

# =========================
# MediaPipe
# =========================
mediapipe==0.10.9

# =========================
# PyTorch CUDA 11.8
# =========================
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.0
torchvision==0.17.0

# =========================
# Utilities
# =========================
python-dotenv==0.21.1
python-json-logger==2.0.4

# =========================
# Dev / test (optional)
# =========================
pytest==7.1.3
pytest-asyncio==0.19.0
httpx==0.23.3
